@@ -1,18 +1,18 @@
 plugins {
-    id "fabric-loom" version "0.5-SNAPSHOT"
+    id "fabric-loom" version "0.9-SNAPSHOT"
     id "maven-publish"
-    id "maven"
+//    id "maven"
 }
 
-sourceCompatibility = JavaVersion.VERSION_1_8
-targetCompatibility = JavaVersion.VERSION_1_8
+sourceCompatibility = JavaVersion.VERSION_16
+targetCompatibility = JavaVersion.VERSION_16
 
 archivesBaseName = project.archives_base_name
 version = project.mod_version
 group = project.maven_group
 
 repositories {
-    maven {url = "https://dl.bintray.com/user11681/maven"}
+    maven {url = "https://auoeke.jfrog.io/artifactory/maven"}
 }
 
 dependencies {
@@ -32,9 +32,9 @@ dependencies {
 //    implementation(include("net.java.dev.jna:jna:5.3.1"))
 //    implementation(include("net.java.dev.jna:jna-platform:5.3.1"))
 
-    modImplementation(include("user11681:dynamicentry:0.0.2"))
-    include(implementation("user11681:reflect:0.2.0"))
-    include(implementation("net.gudenau.lib:unsafe:1.2.14"))
+    modImplementation(include("user11681:dynamic-entry:+"))
+    include(implementation("user11681:reflect:latest.integration"))
+    include(implementation("net.gudenau.lib:unsafe:latest.integration"))
 }
 
 processResources {
@@ -43,18 +43,21 @@ processResources {
     from(sourceSets.main.resources.srcDirs) {
         include "fabric.mod.json"
         expand "version": project.version
+        duplicatesStrategy DuplicatesStrategy.EXCLUDE
     }
 
     from(sourceSets.main.resources.srcDirs) {
         exclude "fabric.mod.json"
+        duplicatesStrategy DuplicatesStrategy.EXCLUDE
     }
 }
 
 // ensure that the encoding is set to UTF-8, no matter what the system default is
 // this fixes some edge cases with special characters not displaying correctly
 // see http://yodaconditions.net/blog/fix-for-java-file-encoding-problems-with-gradle.html
-tasks.withType(JavaCompile) {
-    options.encoding = "UTF-8"
+tasks.withType(JavaCompile).configureEach {
+    it.options.encoding = "UTF-8"
+    it.options.release = 16
 }
 
 // Loom will automatically attach sourcesJar to a RemapSourcesJar task and to the "build" task
