@@ -1,7 +1,5 @@
 package net.devtech.grossfabrichacks.mixin;
 
-import java.util.List;
-import java.util.Set;
 import net.devtech.grossfabrichacks.GrossFabricHacks;
 import net.devtech.grossfabrichacks.entrypoints.PrePreLaunch;
 import net.devtech.grossfabrichacks.transformer.TransformerApi;
@@ -9,6 +7,9 @@ import org.objectweb.asm.tree.ClassNode;
 import org.spongepowered.asm.mixin.extensibility.IMixinConfigPlugin;
 import org.spongepowered.asm.mixin.extensibility.IMixinInfo;
 import user11681.dynamicentry.DynamicEntry;
+
+import java.util.List;
+import java.util.Set;
 
 public class GrossFabricHacksPlugin implements IMixinConfigPlugin {
     @Override
@@ -41,7 +42,7 @@ public class GrossFabricHacksPlugin implements IMixinConfigPlugin {
     static {
         GrossFabricHacks.State.mixinLoaded = true;
 
-        DynamicEntry.executeOptionalEntrypoint("gfh:prePreLaunch", PrePreLaunch.class, PrePreLaunch::onPrePreLaunch);
+        DynamicEntry.execute("gfh:prePreLaunch", PrePreLaunch.class, PrePreLaunch::onPrePreLaunch);
 
         if (GrossFabricHacks.State.shouldWrite || GrossFabricHacks.State.manualLoad) {
             TransformerApi.manualLoad();
