@@ -1,9 +1,11 @@
 package net.devtech.grossfabrichacks.unsafe;
 
+import net.devtech.grossfabrichacks.Rethrower;
+import net.gudenau.lib.unsafe.Unsafe;
+
 import java.io.InputStream;
 import java.lang.reflect.Array;
 import java.security.ProtectionDomain;
-import net.gudenau.lib.unsafe.Unsafe;
 
 /**
  * works across all normal JVMs I think
@@ -174,7 +176,7 @@ public class UnsafeUtil extends Unsafe {
         try {
             putInt(object, objectFieldOffset(object.getClass().getDeclaredField(field)), value);
         } catch (final NoSuchFieldException exception) {
-            throw new RuntimeException(exception);
+            throw Rethrower.rethrow(exception);
         }
     }
 
@@ -182,7 +184,7 @@ public class UnsafeUtil extends Unsafe {
         try {
             putInt(object, objectFieldOffset(klass.getDeclaredField(field)), value);
         } catch (final NoSuchFieldException exception) {
-            throw new RuntimeException(exception);
+            throw Rethrower.rethrow(exception);
         }
     }
 
@@ -255,7 +257,7 @@ public class UnsafeUtil extends Unsafe {
 
             return bytecode;
         } catch (final Throwable throwable) {
-            throw new RuntimeException(throwable);
+            throw Rethrower.rethrow(throwable);
         }
     }
 
@@ -263,7 +265,7 @@ public class UnsafeUtil extends Unsafe {
         try {
             return initialiizeClass(findAndDefineClass(binaryName, loader));
         } catch (final Throwable throwable) {
-            throw new RuntimeException(throwable);
+            throw Rethrower.rethrow(throwable);
         }
     }
 
@@ -271,7 +273,7 @@ public class UnsafeUtil extends Unsafe {
         try {
             return (Class<T>) Class.forName(name);
         } catch (final ClassNotFoundException exception) {
-            throw new RuntimeException(exception);
+            throw Rethrower.rethrow(exception);
         }
     }
 
@@ -279,7 +281,7 @@ public class UnsafeUtil extends Unsafe {
         try {
             return (Class<T>) Class.forName(name, initialize, loader);
         } catch (final ClassNotFoundException exception) {
-            throw new RuntimeException(exception);
+            throw Rethrower.rethrow(exception);
         }
     }
 
@@ -312,7 +314,7 @@ public class UnsafeUtil extends Unsafe {
 
             addressFactor = x64 ? 8 : 1;
         } catch (final Throwable throwable) {
-            throw new RuntimeException(throwable);
+            throw Rethrower.rethrow(throwable);
         }
 
         FIRST_INT_KLASS = getKlass(new FirstInt());
